added exist method

+++ src/main/java/com/ovea/system/util/ProcUtils.java
@@ -97,10 +97,38 @@ public final class ProcUtils {
         }
     }
 
+    public static boolean exist(long pid) {
+        if (Platform.isWindows() || Platform.isWindowsCE()) {
+            WinNT.HANDLE handle = Kernel32.INSTANCE.OpenProcess(0x0400, true, (int) pid);
+            try {
+                return handle != null;
+            } finally {
+                if (handle != null) {
+                    Kernel32.INSTANCE.CloseHandle(handle);
+                }
+            }
+        } else {
+            try {
+                SigarLoader.instance().getProcState(pid);
+                return true;
+            } catch (SigarException e) {
+                return false;
+            }
+        }
+    }
+
     public static void kill(long pid) {
         if (Platform.isWindows() || Platform.isWindowsCE()) {
             WinNT.HANDLE handle = Kernel32.INSTANCE.OpenProcess(0x0001, true, (int) pid);
-            Kernel32.INSTANCE.TerminateProcess(handle, 0);
+            try {
+                if (handle != null) {
+                    Kernel32.INSTANCE.TerminateProcess(handle, 0);
+                }
+            } finally {
+                if (handle != null) {
+                    Kernel32.INSTANCE.CloseHandle(handle);
+                }
+            }
         } else {
             Sigar sigar = SigarLoader.instance();
             try {
